+++ packages/eas-cli/src/graphql/types/Workflow.ts
@@ -1,0 +1,11 @@
+import gql from 'graphql-tag';
+
+export const WorkflowFragmentNode = gql`
+  fragment WorkflowFragment on Workflow {
+    id
+    name
+    fileName
+    createdAt
+    updatedAt
+  }
+`;