+++ packages/eas-cli/src/graphql/types/WorkflowRun.ts
@@ -1,0 +1,17 @@
+import gql from 'graphql-tag';
+
+export const WorkflowRunFragmentNode = gql`
+  fragment WorkflowRunFragment on WorkflowRun {
+    id
+    status
+    gitCommitMessage
+    gitCommitHash
+    createdAt
+    updatedAt
+    workflow {
+      id
+      name
+      fileName
+    }
+  }
+`;